requirements.txt
pyyaml>=5.3.0
paho-mqtt>=1.5.0
requests>=2.22.0